feat: refactor sncrypto to add unified sha256 and base64 usage
