+++ packages/sncrypto-common/src/SHA/SNCryptoSha256Interface.ts
@@ -1,0 +1,6 @@
+import { HexString } from '../Types/HexString'
+import { Utf8String } from '../Types/Utf8String'
+
+export interface SNCryptoSha256Interface {
+  sha256(text: Utf8String): HexString
+}
