+++ packages/sncrypto-common/src/Types/Base64String.ts
@@ -1,0 +1,1 @@
+export type Base64String = string
