+++ packages/sncrypto-common/src/Types/HexString.ts
@@ -1,0 +1,1 @@
+export type HexString = string
