+++ packages/sncrypto-common/src/Types/SodiumConstant.ts
@@ -1,0 +1,11 @@
+export enum SodiumConstant {
+  CRYPTO_SECRETSTREAM_XCHACHA20POLY1305_STATEBYTES = 52,
+  CRYPTO_SECRETSTREAM_XCHACHA20POLY1305_ABYTES = 17,
+  CRYPTO_SECRETSTREAM_XCHACHA20POLY1305_HEADERBYTES = 24,
+  CRYPTO_SECRETSTREAM_XCHACHA20POLY1305_KEYBYTES = 32,
+  CRYPTO_SECRETSTREAM_XCHACHA20POLY1305_TAG_PUSH = 0,
+  CRYPTO_SECRETSTREAM_XCHACHA20POLY1305_TAG_PULL = 1,
+  CRYPTO_SECRETSTREAM_XCHACHA20POLY1305_TAG_REKEY = 2,
+  CRYPTO_SECRETSTREAM_XCHACHA20POLY1305_TAG_FINAL = 3,
+  CRYPTO_SECRETSTREAM_XCHACHA20POLY1305_MESSAGEBYTES_MAX = 0x3fffffff80,
+}
