+++ packages/sncrypto-common/src/Types/SodiumStateAddress.ts
@@ -1,0 +1,1 @@
+export type SodiumStateAddress = unknown
