+++ packages/sncrypto-common/src/Types/StreamDecryptor.ts
@@ -1,0 +1,5 @@
+import { SodiumStateAddress } from './SodiumStateAddress'
+
+export type StreamDecryptor = {
+  state: SodiumStateAddress
+}
