+++ packages/sncrypto-common/src/Types/StreamDecryptorResult.ts
@@ -1,0 +1,6 @@
+import { SodiumConstant } from './SodiumConstant'
+
+export type StreamDecryptorResult = {
+  message: Uint8Array
+  tag: SodiumConstant
+}
