+++ packages/sncrypto-common/src/Types/Utf8String.ts
@@ -1,0 +1,1 @@
+export type Utf8String = string
