+++ packages/sncrypto-node/src/index.ts
@@ -1,1 +1,2 @@
-export { SnCryptoNode } from './crypto'
+export * from './SnCryptoNode'
+export * from './Utils'
