+++ packages/snjs/lib/Services/Files/UseCase/FileDecryptor.ts
@@ -1,10 +1,10 @@
-import { SNPureCrypto, StreamDecryptor, SodiumConstant } from '@standardnotes/sncrypto-common'
+import { SNPureCryptoInterface, StreamDecryptor, SodiumConstant } from '@standardnotes/sncrypto-common'
 import { EncryptedFileInterface } from '../Types'
 
 export class FileDecryptor {
   private decryptor: StreamDecryptor
 
-  constructor(private file: EncryptedFileInterface, private crypto: SNPureCrypto) {
+  constructor(private file: EncryptedFileInterface, private crypto: SNPureCryptoInterface) {
     this.decryptor = this.crypto.xchacha20StreamInitDecryptor(this.file.encryptionHeader, this.file.key)
   }
